refactor: Refactorizando pantalla de login y signup, agregando animaciones
