+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:serviexpress_app/presentation/widgets/login.dart';
+import 'package:serviexpress_app/presentation/pages/auth_screen.dart';
 import 'firebase_options.dart';
 
 void main() {
@@ -14,6 +14,10 @@ class ServiExpressApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(title: 'ServiExpress', home: Scaffold(body: Login(),),);
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'ServiExpress',
+      home: Scaffold(body: AuthScreen()),
+    );
   }
 }
